Allow for a longer description
While this doesn't explicitly state a sentence length of at least one, I
think it can be implied. This gives more leeway for bigger changes.

This brings it more in-line with the original text before any of my
other changes.

Also just elaborated that you're supposed to put a description /of the
changes/, not just a description of whatever.