Remove Probationary Membership
Probationary Membership just seems to fill up the constitution without
actually adding any extra depth to the document.

If a member doesn't register or have issues that impede their
membership, they will simply not be a member. They will have to rejoin
and have to reapply.

Instead, there's going to be more of an emphasis on being able to
re-apply for prior positions (which the only one available is contributing membership).