Remove references to Active Membership
Article 2 (who can be a candidate for executive membership),
Article 6 (who can vote in general elections), and
Article 8 (who can propose amendments)
have been moved to Contributing Membership instead.

The stuff in Article 3 was redundant anyways so I just deleted it.

(Also removed that pesky Probationary Membership thing as it was a
workaround to let members in the next academic year still vote.)