Add Membership Duties
These duties serve as a way to clarify the job each Membership Role
should play.

Contributing Members should have some goal/desire to contribute, and
should usually have a specific thing which they want to accomplish.
Their duties, thus, become accomplishing this desire. If they want to
contribute to our open source projects, have them do that.

Clarified that Technology is a privilege, not a right. Not everyone will
have access to the server, and that's by design. If you don't need it or
can't make a case for using it, you don't get to use it.

I removed the resignation thing because it's just unnecessary and
unenforceable. Just give notice and leave.

Because probationary membership got removed, it just makes sense to
allow them to simply request membership once again at the start of the
year and it to be leniently granted on-request. It should be easier to
get Contributing Membership through this than through the normal way as
your desire is already shown and you are known to be helpful.

The Executive Board duties just reflect the general idea that they
should be stewards of the organization and should be doing everything
good.

Probational Membership for Executive Membership got removed entirely and
replaced by getting membership in the next year, as elections will be
moved to the end of the year.