Move Elections to end of year
To aid in removing probational membership, it makes sense to just move
elections to the end of the year.

Thus, there is time for new members to become active in the club and
gain privileges to vote and allow for out-going students to shape the
future of the club.

Also it's just easier because everyone else does it.

The 6 week thing is pretty arbitrary though, so we can change that whenever.

Also removed a dangling Active Membership reference :).